+++ Sources/SwiftSDK/Messaging/DeliveryOptions.swift
@@ -34,20 +34,67 @@
     case FOR_ALL = 15
 }
 
-@objcMembers open class DeliveryOptions: NSObject {
+@objcMembers open class DeliveryOptions: NSObject, NSCoding, Codable {
     
     open var publishAt: Date?
-    open var repeatEvery: NSNumber?
     open var repeatExpiresAt: Date?
+
+    private var _repeatEvery: Int?
+    open var repeatEvery: NSNumber? {
+        get {
+            if let _repeatEvery = _repeatEvery {
+                return NSNumber(integerLiteral: _repeatEvery)
+            }
+            return nil
+        }
+        set(newRepeatEvery) {
+            _repeatEvery = newRepeatEvery?.intValue
+        }
+    }
+
+    private var pushSinglecast: [String] = []
+    private var publishPolicy: Int = PublishPolicyEnum.BOTH.rawValue
+    private var pushBroadcast: Int = PushBroadcastEnum.FOR_ALL.rawValue
     
-    private var pushSinglecast: [String]!
-    private var publishPolicy: NSNumber!
-    private var pushBroadcast: NSNumber!
+    enum CodingKeys: String, CodingKey {
+        case publishAt
+        case repeatExpiresAt
+        case _repeatEvery = "repeatEvery"
+        case pushSinglecast
+        case publishPolicy
+        case pushBroadcast
+    }
     
-    public override init() {
-        self.pushSinglecast = [String]()
-        self.publishPolicy = NSNumber(integerLiteral: PublishPolicyEnum.BOTH.rawValue)
-        self.pushBroadcast = NSNumber(integerLiteral: PushBroadcastEnum.FOR_ALL.rawValue)
+    public override init() { }
+    
+    public init(publishAt: Date?, repeatExpiresAt: Date?, _repeatEvery: Int?, pushSinglecast: [String]?, publishPolicy: Int, pushBroadcast: Int) {
+        self.publishAt = publishAt
+        self.repeatExpiresAt = repeatExpiresAt
+        self._repeatEvery = _repeatEvery
+        if pushSinglecast != nil {
+            self.pushSinglecast = pushSinglecast!
+        }
+        self.publishPolicy = publishPolicy
+        self.pushBroadcast = pushBroadcast
+    }
+    
+    convenience public required init?(coder aDecoder: NSCoder) {
+        let publishAt = aDecoder.decodeObject(forKey: CodingKeys.publishAt.rawValue) as? Date
+        let repeatExpiresAt = aDecoder.decodeObject(forKey: CodingKeys.repeatExpiresAt.rawValue) as? Date
+        let _repeatEvery = aDecoder.decodeInteger(forKey: CodingKeys._repeatEvery.rawValue)
+        let pushSinglecast = aDecoder.decodeObject(forKey: CodingKeys.pushSinglecast.rawValue) as? [String]
+        let publishPolicy = aDecoder.decodeInteger(forKey: CodingKeys.publishPolicy.rawValue)
+        let pushBroadcast = aDecoder.decodeInteger(forKey: CodingKeys.pushBroadcast.rawValue)
+        self.init(publishAt: publishAt, repeatExpiresAt: repeatExpiresAt, _repeatEvery: _repeatEvery, pushSinglecast: pushSinglecast, publishPolicy: publishPolicy, pushBroadcast: pushBroadcast)
+    }
+    
+    public func encode(with aCoder: NSCoder) {
+        aCoder.encode(publishAt, forKey: CodingKeys.publishAt.rawValue)
+        aCoder.encode(repeatExpiresAt, forKey: CodingKeys.repeatExpiresAt.rawValue)
+        aCoder.encode(_repeatEvery, forKey: CodingKeys._repeatEvery.rawValue)
+        aCoder.encode(pushSinglecast, forKey: CodingKeys.pushSinglecast.rawValue)
+        aCoder.encode(publishPolicy, forKey: CodingKeys.publishPolicy.rawValue)
+        aCoder.encode(pushBroadcast, forKey: CodingKeys.pushBroadcast.rawValue)
     }
     
     open func setPushSinglecast(singlecast: [String]) {
@@ -55,12 +102,12 @@
     }
     
     open func addPushSingleCast(singlecast: String) {
-        self.pushSinglecast?.append(singlecast)
+        self.pushSinglecast.append(singlecast)
     }
     
     open func removePushSinglecast(singlecast: String) {
-        if let index = self.pushSinglecast?.firstIndex(of: singlecast) {
-            self.pushSinglecast?.remove(at: index)
+        if let index = self.pushSinglecast.firstIndex(of: singlecast) {
+            self.pushSinglecast.remove(at: index)
         }
     }
     
@@ -69,7 +116,7 @@
     }
     
     open func setPublishPolicy(publishPolicy: Int) {
-        self.publishPolicy = NSNumber(integerLiteral: publishPolicy)
+        self.publishPolicy = publishPolicy
     }
     
     open func getPublishPolicy() -> String {
@@ -83,10 +130,10 @@
     }
     
     open func setPushBroadcast(pushBroadcast: Int) {
-        self.pushBroadcast = NSNumber(integerLiteral: pushBroadcast)
+        self.pushBroadcast = pushBroadcast
     }
     
     open func getPushBroadcast() -> Int {
-        return self.pushBroadcast.intValue
+        return self.pushBroadcast
     }
 }